Move search suggestion client/model to a common place
