+++ catowseriOS/catowser/Search/SearchBarViewModel.swift
@@ -30,26 +30,6 @@ final class SearchBarViewModel: NSObject, ObservableObject {
     /// Can't declare it private due to compiler error.
     @LeadingTrimmed private var tempSearchText: String
     
-    private nonisolated var searchSuggestClient: SearchEngine {
-        get async {
-            let name = await FeatureManager.shared.searchPluginName()
-            let optionalXmlData = ResourceReader.readXmlSearchPlugin(with: name, on: .main)
-            guard let xmlData = optionalXmlData else {
-                return .googleSearchEngine()
-            }
-            
-            let osDescription: OpenSearch.Description
-            do {
-                osDescription = try OpenSearch.Description(data: xmlData)
-            } catch {
-                print("Open search xml parser error: \(error.localizedDescription)")
-                return .googleSearchEngine()
-            }
-            
-            return osDescription.html
-        }
-    }
-    
     override init() {
         showSearchSuggestions = false
         searchQuery = ""
@@ -93,7 +73,8 @@ extension SearchBarViewModel: SearchSuggestionsListDelegate {
                 }
                 replaceTab(with: url, with: nil, isJSEnabled)
             case .suggestion(let suggestion):
-                guard let url = await searchSuggestClient.searchURLForQuery(suggestion) else {
+                let client = await HttpEnvironment.shared.searchSuggestClient()
+                guard let url = client.searchURLForQuery(suggestion) else {
                     assertionFailure("Failed construct search engine url from suggestion string")
                     return
                 }
